feat: Add show method to NotificationController
This commit adds a new `show` method to the `NotificationController.php` file. The method retrieves a notification by its ID, marks it as read, and redirects the user based on the notification type. If the type is 'reservation_client', the user is redirected to the reservations index page. If the type is 'reservation_gestionnaire', the user is redirected to the gestionnaire reservations requests page.